+++ src/main/java/kitchenpos/menus/application/MenuGroupService.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.application;
+package kitchenpos.menus.application;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
