+++ src/main/java/kitchenpos/menus/domain/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menus.domain;
 
 import java.util.List;
 import java.util.Optional;
