+++ src/main/java/kitchenpos/menus/tobe/domain/JpaMenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menus.tobe.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
