+++ src/main/java/kitchenpos/menus/tobe/domain/JpaMenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menus.tobe.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
