+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,8 +1,8 @@
 package kitchenpos.order.application;
 
 import kitchenpos.infra.KitchenridersClient;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.order.domain.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
