+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.domain;
 
 import jakarta.persistence.*;
-import kitchenpos.menu.domain.Menu;
+import kitchenpos.menus.domain.Menu;
 
 import java.math.BigDecimal;
 import java.util.UUID;
