+++ src/test/java/kitchenpos/fixture/MenuProductFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menus.domain.MenuProduct;
 import kitchenpos.products.tobe.domain.Product;
 import org.jetbrains.annotations.NotNull;
