+++ src/test/java/kitchenpos/menu/domain/InMemoryMenuGroupRepository.java
@@ -1,5 +1,8 @@
 package kitchenpos.menu.domain;
 
+import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuGroupRepository;
+
 import java.util.*;
 
 public class InMemoryMenuGroupRepository implements MenuGroupRepository {
