+++ src/test/java/kitchenpos/menu/tobe/domain/MenuPriceTest.java
@@ -1,6 +1,7 @@
 package kitchenpos.menu.tobe.domain;
 
 import kitchenpos.exception.IllegalPriceException;
+import kitchenpos.menus.tobe.domain.MenuPrice;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
