+++ src/test/java/kitchenpos/menu/tobe/domain/MenuTest.java
@@ -4,6 +4,7 @@ import kitchenpos.exception.IllegalPriceException;
 import kitchenpos.fixture.tobe.ProductFixture;
 import kitchenpos.infra.FakePurgomalumClient;
 import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.menus.tobe.domain.*;
 import kitchenpos.products.tobe.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
